// projects/hr-smart/src/app/webservice/generated/hr-smart-service/model/participantEntry.ts
/**
 * HR-Smart Logging-Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TrainingEntry } from './trainingEntry';
import { UserEntry } from './userEntry';


export interface ParticipantEntry { 
    participantId?: string;
    training?: TrainingEntry;
    trainings?: UserEntry;
}